Move secret/configmap automount code to separate file
Move functions related to automounting configmaps and secrets to a
separate file to make reading deployment.go easier.
